+++ src/DataBlobs/services/BlobStores.jl
@@ -6,7 +6,7 @@
 Get the data blob for the specified blobstore or dfg.
 
 Related
-[`getBlobEntry`](@ref)
+[`getBlobentry`](@ref)
 
 $(METHODLIST)
 """
@@ -16,7 +16,7 @@ function getBlob end
 Adds a blob to the blob store or dfg with the given entry.
 
 Related
-[`addBlobEntry!`](@ref)
+[`addBlobentry!`](@ref)
 
 $(METHODLIST)
 """
@@ -38,7 +38,7 @@ function updateBlob! end
 Delete a blob from the blob store or dfg with the given entry.
 
 Related
-[`deleteBlobEntry!`](@ref)
+[`deleteBlobentry!`](@ref)
 
 $(METHODLIST)
 """
@@ -51,39 +51,39 @@ List all ids in the blob store.
 function listBlobs end
 
 ##==============================================================================
-## AbstractBlobStore CRUD Interface
+## AbstractBlobstore CRUD Interface
 ##==============================================================================
 
-function getBlob(store::AbstractBlobStore, ::UUID)
+function getBlob(store::AbstractBlobstore, ::UUID)
     return error("$(typeof(store)) doesn't override 'getBlob'.")
 end
 
-function addBlob!(store::AbstractBlobStore{T}, ::UUID, ::T) where {T}
+function addBlob!(store::AbstractBlobstore{T}, ::UUID, ::T) where {T}
     return error("$(typeof(store)) doesn't override 'addBlob!'.")
 end
 
-function updateBlob!(store::AbstractBlobStore{T}, ::UUID, ::T) where {T}
+function updateBlob!(store::AbstractBlobstore{T}, ::UUID, ::T) where {T}
     return error("$(typeof(store)) doesn't override 'updateBlob!'.")
 end
 
-function deleteBlob!(store::AbstractBlobStore, ::UUID)
+function deleteBlob!(store::AbstractBlobstore, ::UUID)
     return error("$(typeof(store)) doesn't override 'deleteBlob!'.")
 end
 
-function listBlobs(store::AbstractBlobStore)
+function listBlobs(store::AbstractBlobstore)
     return error("$(typeof(store)) doesn't override 'listBlobs'.")
 end
 
-function hasBlob(store::AbstractBlobStore, ::UUID)
+function hasBlob(store::AbstractBlobstore, ::UUID)
     return error("$(typeof(store)) doesn't override 'hasBlob'.")
 end
 
 ##==============================================================================
-## AbstractBlobStore derived CRUD for Blob 
+## AbstractBlobstore derived CRUD for Blob 
 ##==============================================================================
 
-function getBlob(dfg::AbstractDFG, entry::BlobEntry)
-    stores = getBlobStores(dfg)
+function getBlob(dfg::AbstractDFG, entry::Blobentry)
+    stores = getBlobstores(dfg)
     storekeys = collect(keys(stores))
     # first check the saved blobstore and then fall back to the rest
     fidx = findfirst(==(entry.blobstore), storekeys)
@@ -110,45 +110,45 @@ function getBlob(dfg::AbstractDFG, entry::BlobEntry)
     )
 end
 
-function getBlob(store::AbstractBlobStore, entry::BlobEntry)
+function getBlob(store::AbstractBlobstore, entry::Blobentry)
     blobId = isnothing(entry.blobId) ? entry.originId : entry.blobId
     return getBlob(store, blobId)
 end
 
 #add 
-function addBlob!(dfg::AbstractDFG, entry::BlobEntry, data)
-    return addBlob!(getBlobStore(dfg, entry.blobstore), entry, data)
+function addBlob!(dfg::AbstractDFG, entry::Blobentry, data)
+    return addBlob!(getBlobstore(dfg, entry.blobstore), entry, data)
 end
 
-function addBlob!(store::AbstractBlobStore{T}, entry::BlobEntry, data::T) where {T}
+function addBlob!(store::AbstractBlobstore{T}, entry::Blobentry, data::T) where {T}
     blobId = isnothing(entry.blobId) ? entry.originId : entry.blobId
     return addBlob!(store, blobId, data)
 end
 
 # also creates an originId as uuid4
-addBlob!(store::AbstractBlobStore, data) = addBlob!(store, uuid4(), data)
+addBlob!(store::AbstractBlobstore, data) = addBlob!(store, uuid4(), data)
 
 #update
-function updateBlob!(dfg::AbstractDFG, entry::BlobEntry, data)
-    return updateBlob!(getBlobStore(dfg, entry.blobstore), entry.blobId, data)
+function updateBlob!(dfg::AbstractDFG, entry::Blobentry, data)
+    return updateBlob!(getBlobstore(dfg, entry.blobstore), entry.blobId, data)
 end
 
-function updateBlob!(store::AbstractBlobStore, entry::BlobEntry, data)
+function updateBlob!(store::AbstractBlobstore, entry::Blobentry, data)
     return updateBlob!(store, entry.blobId, data)
 end
 #delete
-function deleteBlob!(dfg::AbstractDFG, entry::BlobEntry)
-    return deleteBlob!(getBlobStore(dfg, entry.blobstore), entry)
+function deleteBlob!(dfg::AbstractDFG, entry::Blobentry)
+    return deleteBlob!(getBlobstore(dfg, entry.blobstore), entry)
 end
 
-function deleteBlob!(store::AbstractBlobStore, entry::BlobEntry)
+function deleteBlob!(store::AbstractBlobstore, entry::Blobentry)
     blobId = isnothing(entry.blobId) ? entry.originId : entry.blobId
     return deleteBlob!(store, blobId)
 end
 
 #has
-function hasBlob(dfg::AbstractDFG, entry::BlobEntry)
-    return hasBlob(getBlobStore(dfg, entry.blobstore), entry.originId)
+function hasBlob(dfg::AbstractDFG, entry::Blobentry)
+    return hasBlob(getBlobstore(dfg, entry.blobstore), entry.originId)
 end
 
 #TODO
@@ -158,7 +158,7 @@ end
 # Can specify which entries to copy with the `sourceEntries` parameter.
 # Returns the list of copied entries.
 # """
-# function copyBlobStore(sourceStore::D1, destStore::D2; sourceEntries=listEntries(sourceStore))::Vector{E} where {T, D1 <: AbstractDataStore{T}, D2 <: AbstractDataStore{T}, E <: BlobEntry}
+# function copyBlobstore(sourceStore::D1, destStore::D2; sourceEntries=listEntries(sourceStore))::Vector{E} where {T, D1 <: AbstractDataStore{T}, D2 <: AbstractDataStore{T}, E <: Blobentry}
 #     # Quick check
 #     destEntries = listBlobs(destStore)
 #     typeof(sourceEntries) != typeof(destEntries) && error("Can't copy stores, source has entries of type $(typeof(sourceEntries)), destination has entries of type $(typeof(destEntries)).")
@@ -174,7 +174,7 @@ end
 ##==============================================================================
 ## FolderStore
 ##==============================================================================
-struct FolderStore{T} <: AbstractBlobStore{T}
+struct FolderStore{T} <: AbstractBlobstore{T}
     label::Symbol
     folder::String
 end
@@ -241,30 +241,30 @@ function hasBlob(store::FolderStore, blobId::UUID)
     return isfile(blobfilename)
 end
 
-hasBlob(store::FolderStore, entry::BlobEntry) = hasBlob(store, entry.originId)
+hasBlob(store::FolderStore, entry::Blobentry) = hasBlob(store, entry.originId)
 
 listBlobs(store::FolderStore) = readdir(store.folder)
 ##==============================================================================
-## InMemoryBlobStore
+## InMemoryBlobstore
 ##==============================================================================
 
-struct InMemoryBlobStore{T} <: AbstractBlobStore{T}
+struct InMemoryBlobstore{T} <: AbstractBlobstore{T}
     label::Symbol
     blobs::Dict{UUID, T}
 end
 
-function InMemoryBlobStore{T}(storeKey::Symbol) where {T}
-    return InMemoryBlobStore{T}(storeKey, Dict{UUID, T}())
+function InMemoryBlobstore{T}(storeKey::Symbol) where {T}
+    return InMemoryBlobstore{T}(storeKey, Dict{UUID, T}())
 end
-function InMemoryBlobStore(storeKey::Symbol = :default_inmemory_store)
-    return InMemoryBlobStore{Vector{UInt8}}(storeKey)
+function InMemoryBlobstore(storeKey::Symbol = :default_inmemory_store)
+    return InMemoryBlobstore{Vector{UInt8}}(storeKey)
 end
 
-function getBlob(store::InMemoryBlobStore, blobId::UUID)
+function getBlob(store::InMemoryBlobstore, blobId::UUID)
     return store.blobs[blobId]
 end
 
-function addBlob!(store::InMemoryBlobStore{T}, blobId::UUID, data::T) where {T}
+function addBlob!(store::InMemoryBlobstore{T}, blobId::UUID, data::T) where {T}
     if haskey(store.blobs, blobId)
         error("Key '$blobId' blob already exists.")
     end
@@ -272,27 +272,27 @@ function addBlob!(store::InMemoryBlobStore{T}, blobId::UUID, data::T) where {T}
     return blobId
 end
 
-function updateBlob!(store::InMemoryBlobStore{T}, blobId::UUID, data::T) where {T}
+function updateBlob!(store::InMemoryBlobstore{T}, blobId::UUID, data::T) where {T}
     if haskey(store.blobs, blobId)
         @warn "Key '$blobId' doesn't exist."
     end
     return store.blobs[blobId] = data
 end
 
-function deleteBlob!(store::InMemoryBlobStore, blobId::UUID)
+function deleteBlob!(store::InMemoryBlobstore, blobId::UUID)
     pop!(store.blobs, blobId)
     return 1
 end
 
-hasBlob(store::InMemoryBlobStore, blobId::UUID) = haskey(store.blobs, blobId)
+hasBlob(store::InMemoryBlobstore, blobId::UUID) = haskey(store.blobs, blobId)
 
-listBlobs(store::InMemoryBlobStore) = collect(keys(store.blobs))
+listBlobs(store::InMemoryBlobstore) = collect(keys(store.blobs))
 
 ##==============================================================================
 ## LinkStore Link blobId to a existing local folder
 ##==============================================================================
 
-struct LinkStore <: AbstractBlobStore{String}
+struct LinkStore <: AbstractBlobstore{String}
     label::Symbol
     csvfile::String
     cache::Dict{UUID, String}
@@ -318,7 +318,7 @@ function getBlob(store::LinkStore, blobId::UUID)
     return read(fname)
 end
 
-function addBlob!(store::LinkStore, entry::BlobEntry, linkfile::String)
+function addBlob!(store::LinkStore, entry::Blobentry, linkfile::String)
     return addBlob!(store, entry.originId, nothing, linkfile::String)
 end
 
@@ -337,11 +337,11 @@ function deleteBlob!(store::LinkStore, args...)
     return error("deleteDataBlob(::LinkStore) not supported")
 end
 
-deleteBlob!(store::LinkStore, ::BlobEntry) = deleteBlob!(store)
+deleteBlob!(store::LinkStore, ::Blobentry) = deleteBlob!(store)
 deleteBlob!(store::LinkStore, ::UUID) = deleteBlob!(store)
 
 ##==============================================================================
-## RowBlobStore Ordered Dict Row Table Blob Store
+## RowBlobstore Ordered Dict Row Table Blob Store
 ##==============================================================================
 
 # RowBlob
@@ -368,22 +368,22 @@ function Tables.columnnames(row::RowBlob)
     return (:id, Tables.columnnames(getfield(row, :blob))...)
 end
 
-## RowBlobStore
+## RowBlobstore
 
-struct RowBlobStore{T} <: AbstractBlobStore{T}
+struct RowBlobstore{T} <: AbstractBlobstore{T}
     label::Symbol
     blobs::OrderedDict{UUID, RowBlob{T}}
 end
 
-function RowBlobStore{T}(storeKey::Symbol) where {T}
-    return RowBlobStore{T}(storeKey, OrderedDict{UUID, RowBlob{T}}())
+function RowBlobstore{T}(storeKey::Symbol) where {T}
+    return RowBlobstore{T}(storeKey, OrderedDict{UUID, RowBlob{T}}())
 end
-function RowBlobStore(storeKey::Symbol, T::DataType)
-    return RowBlobStore{T}(storeKey)
+function RowBlobstore(storeKey::Symbol, T::DataType)
+    return RowBlobstore{T}(storeKey)
 end
 
-function RowBlobStore(storeKey::Symbol, T::DataType, table)
-    store = RowBlobStore(storeKey, T)
+function RowBlobstore(storeKey::Symbol, T::DataType, table)
+    store = RowBlobstore(storeKey, T)
     for nt in Tables.namedtupleiterator(table)
         row = DFG.RowBlob(T, nt)
         store.blobs[row.id] = row
@@ -392,18 +392,18 @@ function RowBlobStore(storeKey::Symbol, T::DataType, table)
 end
 
 # Tables interface
-Tables.istable(::Type{RowBlobStore{T}}) where {T} = true
-Tables.rowaccess(::Type{RowBlobStore{T}}) where {T} = true
-Tables.rows(store::RowBlobStore) = values(store.blobs)
+Tables.istable(::Type{RowBlobstore{T}}) where {T} = true
+Tables.rowaccess(::Type{RowBlobstore{T}}) where {T} = true
+Tables.rows(store::RowBlobstore) = values(store.blobs)
 #TODO
-# Tables.materializer(::Type{RowBlobStore{T}}) where T = Tables.rowtable
+# Tables.materializer(::Type{RowBlobstore{T}}) where T = Tables.rowtable
 
 ##
-function getBlob(store::RowBlobStore, blobId::UUID)
+function getBlob(store::RowBlobstore, blobId::UUID)
     return getfield(store.blobs[blobId], :blob)
 end
 
-function addBlob!(store::RowBlobStore{T}, blobId::UUID, blob::T) where {T}
+function addBlob!(store::RowBlobstore{T}, blobId::UUID, blob::T) where {T}
     if haskey(store.blobs, blobId)
         error("Key '$blobId' blob already exists.")
     end
@@ -411,21 +411,21 @@ function addBlob!(store::RowBlobStore{T}, blobId::UUID, blob::T) where {T}
     return blobId
 end
 
-function updateBlob!(store::RowBlobStore{T}, blobId::UUID, blob::T) where {T}
+function updateBlob!(store::RowBlobstore{T}, blobId::UUID, blob::T) where {T}
     if haskey(store.blobs, blobId)
         @warn "Key '$blobId' doesn't exist."
     end
     return store.blobs[blobId] = RowBlob(blobId, blob)
 end
 
-function deleteBlob!(store::RowBlobStore, blobId::UUID)
+function deleteBlob!(store::RowBlobstore, blobId::UUID)
     getfield(pop!(store.blobs, blobId), :blob)
     return 1
 end
 
-hasBlob(store::RowBlobStore, blobId::UUID) = haskey(store.blobs, blobId)
+hasBlob(store::RowBlobstore, blobId::UUID) = haskey(store.blobs, blobId)
 
-listBlobs(store::RowBlobStore) = collect(keys(store.blobs))
+listBlobs(store::RowBlobstore) = collect(keys(store.blobs))
 
 # TODO also see about wrapping a table directly
 ##
@@ -434,7 +434,7 @@ if false
 
     Tables.columnnames(rb)
 
-    tstore = RowBlobStore(:namedtuple, @NamedTuple{a::Vector{Int}, b::Vector{Int}})
+    tstore = RowBlobstore(:namedtuple, @NamedTuple{a::Vector{Int}, b::Vector{Int}})
 
     addBlob!(tstore, uuid4(), (a = [1, 2], b = [3, 4]))
     addBlob!(tstore, uuid4(), (a = [5, 6], b = [7, 8]))
@@ -456,7 +456,7 @@ if false
         b::Float64
     end
 
-    sstore = RowBlobStore(:struct_Foo, Foo)
+    sstore = RowBlobstore(:struct_Foo, Foo)
 
     addBlob!(sstore, uuid4(), Foo(1, 2))
     addBlob!(sstore, uuid4(), Foo(3, 4))